Fix image picker crash on IOS (#1051)
Fix ImagePicker Crash on IOS by checking safe guarding _result to be not nil.

The method dismissViewControllerAnimated does not immediately prevent further
didFinishPickingMediaWithInfo invocations. So the code in the delegate method get triggered multiple times if user taps multiple times. 

The crash is caused because the _result block is nil and gets called again. By safe guarding it, the code will only be triggered once and will not cause crash.

Another way to fix the crash would be put the code handling selection inside the completion handler inside the dismissViewControllerAnimated:completion method. This will cause a bit delay on user experience since the completion handler is called after dismissing view controller animation finishes. So the fix provided in this PR is favored.