[webview_flutter] Fix double tap on resize issue (#1965)
WebView is incorrectly dropping some input events as unhandled.
Previously we were restarting the input connection to try and work
around this, but that causes a long tail of bugs in various SDK and
WebView versions. Stop the events from bubbling instead since they still
appear to be correctly handled despite being marked wrong.

Long term we should figure out why this is happening in WebView and
figure out a deeper fix for this issue, but this workaround solves the
linked bugs in the immediate short term.