[image_picker] Don't store null paths in lost cache (#6678)
If the user cancels image selection on Android, store nothing in the
lost image cache rather than storing an array with a null path.

While we could potentially keep this behavior and instead handle it
differently on the Dart side, returning some new "cancelled"
`LostDataResponse`, that would be semi-breaking; e.g., the current
example's lost data handling would actually throw as written if we had a
new non-`isEmpty`, non-exception, null-`file` response. Since nobody has
requested the ability to specifically detect a "lost cancel" as being
different from not having started the process of picking anything, this
doesn't make that potentially-client-breaking change. If it turns out
there's a use case for that in the future, we can revisit that (but
should not do it by storing a null entry in a file array anyway).

Fixes https://github.com/flutter/flutter/issues/114551